requirements.txt
# UI Framework
kivymd==1.1.1
kivy==2.2.1

# Telemetry Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6

# Hardware Communication
pymodbus==3.5.4
stupidArtnet==1.4.0
RPi.GPIO==0.7.1; platform_machine == "aarch64" or platform_machine == "armv7l"

# Audio
pygame==2.5.2

# Configuration & Data
pyyaml==6.0.1
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
aiosqlite==0.19.0

# Utilities
python-dotenv==1.0.0
httpx==0.25.2

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
